requirements.txt
ddgs>=9.10.0
duckduckgo-search>=8.1.1
fastmcp>=2.14.3
google-api-core>=2.29.0
google-auth>=2.47.0
google-cloud-aiplatform>=1.133.0
google-cloud-bigquery>=3.40.0
google-cloud-core>=2.5.0
google-cloud-resource-manager>=1.15.0
google-cloud-storage>=2.19.0
google-crc32c>=1.8.0
google-genai>=1.57.0
google-resumable-media>=2.8.0
googleapis-common-protos>=1.72.0
grpc-google-iam-v1>=0.14.3
grpcio>=1.76.0
grpcio-status>=1.76.0
h11>=0.16.0
h2>=4.3.0
hpack>=4.1.0
httpcore>=1.0.9
httpx>=0.28.1
httpx-sse>=0.4.3
hyperframe>=6.1.0
idna>=3.11
importlib-metadata>=8.7.1
ipykernel>=7.1.0
ipython>=9.9.0
ipython-pygments-lexers>=1.1.1
jaraco-classes>=3.4.0
jaraco-context>=6.1.0
jaraco-functools>=4.4.0
jedi>=0.19.2
jiter>=0.12.0
jsonpatch>=1.33
jsonpointer>=3.0.0
jsonschema>=4.26.0
jsonschema-path>=0.3.4
jsonschema-specifications>=2025.9.1
jupyter-client>=8.8.0
jupyter-core>=5.9.1
keyring>=25.7.0
langchain>=1.2.3
langchain-classic>=1.0.1
langchain-community>=0.3.31
langchain-core>=1.2.7
langchain-google-vertexai>=2.1.2
langchain-mcp>=0.2.1
langchain-openai>=0.3.35
langchain-qdrant>=1.1.0
langchain-text-splitters>=0.3.11
langgraph>=1.0.5
langgraph-checkpoint>=3.0.1
langgraph-prebuilt>=1.0.5
langgraph-sdk>=0.3.2
langsmith>=0.6.2
lupa>=2.6
lxml>=6.0.2
markdown-it-py>=4.0.0
marshmallow>=3.26.2
matplotlib-inline>=0.2.1
mcp>=1.25.0
mdurl>=0.1.2
more-itertools>=10.8.0
multidict>=6.7.0
mypy-extensions>=1.1.0
nest-asyncio>=1.6.0
numexpr>=2.14.1
numpy>=2.4.1
openai>=2.15.0
openapi-pydantic>=0.5.1
opentelemetry-api>=1.39.1
opentelemetry-exporter-prometheus>=0.60b1
opentelemetry-instrumentation>=0.60b1
opentelemetry-sdk>=1.39.1
opentelemetry-semantic-conventions>=0.60b1
orjson>=3.11.5
ormsgpack>=1.12.1
packaging>=25.0
pandas>=2.3.3
parso>=0.8.5
pathable>=0.4.4
pathvalidate>=3.3.1
pexpect>=4.9.0
platformdirs>=4.5.1
portalocker>=3.2.0
primp>=0.15.0
prometheus-client>=0.24.1
prompt-toolkit>=3.0.52
propcache>=0.4.1
proto-plus>=1.27.0
protobuf>=6.33.3
psutil>=7.2.1
ptyprocess>=0.7.0
pure-eval>=0.2.3
py-key-value-aio>=0.3.0
py-key-value-shared>=0.3.0
pyarrow>=21.0.0
pyasn1>=0.6.1
pyasn1-modules>=0.4.2
pycparser>=2.23
pydantic>=2.12.5
pydantic-core>=2.41.5
pydantic-settings>=2.12.0
pydocket>=0.16.6
pygments>=2.19.2
pyjwt>=2.10.1
pyperclip>=1.11.0
python-dateutil>=2.9.0.post0
python-dotenv>=1.2.1
python-json-logger>=4.0.0
python-multipart>=0.0.21
pytz>=2025.2
pyyaml>=6.0.3
pyzmq>=27.1.0
qdrant-client>=1.16.2
redis>=7.1.0
referencing>=0.36.2
regex>=2025.11.3
requests>=2.32.5
requests-toolbelt>=1.0.0
rich>=14.2.0
rich-rst>=1.3.2
rpds-py>=0.30.0
rsa>=4.9.1
shellingham>=1.5.4
six>=1.17.0
sniffio>=1.3.1
socksio>=1.0.0
sortedcontainers>=2.4.0
sqlalchemy>=2.0.45
sse-starlette>=3.1.2
stack-data>=0.6.3
starlette>=0.51.0
tabulate>=0.9.0
tenacity>=9.1.2
tiktoken>=0.12.0
tornado>=6.5.4
tqdm>=4.67.1
traitlets>=5.14.3
typer>=0.21.1
typing-extensions>=4.15.0
typing-inspect>=0.9.0
typing-inspection>=0.4.2
tzdata>=2025.3
urllib3>=2.6.3
uuid-utils>=0.13.0
uvicorn>=0.40.0
validators>=0.35.0
wcwidth>=0.2.14
websockets>=15.0.1
wrapt>=1.17.3
xxhash>=3.6.0
yarl>=1.22.0
zipp>=3.23.0
zstandard>=0.25.0